### 2023-10-19 #5962 @editorialbot
[PRE REVIEW]: atems: Analysis tools for TEM images of carbonaceous particles
**Submitting author:** @tsipkens (<a href="http://orcid.org/0000-0003-1719-7105">Timothy A. Sipkens</a>)
**Repository:** https://github.com/tsipkens/atems
**Branch with paper.md** (empty if default branch):
**Version:** v1.2
