### 2023-10-31 #5962 @tsipkens
Excellent! @diazrenata, I will give this some further thought. I will also incorporate the other changes suggested in this thread to make the format more suitable for JOSS.
