### 2023-10-31 076f53b
Add sharpness to analyze_binary
+ A binner function.
