### 2023-11-09 d655870
Bug fix to analyze_binary depth
