### 2023-12-02 #5962 @ajinkyakadu
Hi Renata:

Sorry for the late reply. I'm currently overcommitted with several projects
and wouldn't be able to dedicate the necessary time and attention to review
this submission to JOSS effectively.

Best,
Ajinkya

On Wed, Nov 29, 2023 at 9:01 PM Renata Diaz ***@***.***>
wrote:

> Hi @jonbmartin <https://github.com/jonbmartin>, @ajinkyakadu
> <https://github.com/ajinkyakadu> - would you be available and interested
> in reviewing this submission to JOSS?
>
