### 2023-12-02 #5962 @jonbmartin
Hi Renata,
I should be able to review this article if you still need a reviewer.
Please let me know.
Jonathan

On Wed, Nov 29, 2023 at 2:01 PM Renata Diaz ***@***.***>
wrote:

> Hi @jonbmartin <https://github.com/jonbmartin>, @ajinkyakadu
> <https://github.com/ajinkyakadu> - would you be available and interested
> in reviewing this submission to JOSS?
>
