### 2023-12-05 #5962 @diazrenata
@rickhelmus - would you be available and interested in reviewing this submission to JOSS?
