### 2023-12-14 362d181
Create pair_corr.m
A function to calculate the density-density correlation function. Uses Monte Carlo sampling over a vector v.
