### 2023-12-14 #5962 @diazrenata
@lcmpikg, would you be able and amenable to reviewing this submission to JOSS?
