### 2024-01-05 #5962 @diazrenata
@jonatanalvelid, would you be able and interested in reviewing this submission to JOSS?  I ask particularly because it's a MATLAB submission for image processing of microscopy.
