### 2024-01-11 7d9c8f3
Update load_imgs with another import
Look for black footer and do OCR.
