### 2024-01-12 #5962 @diazrenata
@mdpetters - would you be available and interested in reviewing this submission to JOSS?

(I believe I emailed you with a review invitation back in November - I'm not sure if I reached the correct address, so I'm reaching out via GitHub. I apologize if this is redundant!)
