### 2024-01-20 10cbc9e
Update load_imgs
+ Better read NRC images.
+ Better comments.
