+++ +tools/load_imgs.m
@@ -161,14 +161,14 @@ for jj=1:length(Imgs)
     
     footer_found = 0; % flag whether footer was found
     
-    f_nm = 1;  % flag indicating nanometers (detected below)
+    fl_nm = 1;  % flag indicating nanometers (detected below)
     
     % Search for row satisying 
     f_footrow = sum(Imgs(jj).raw, 2) > ...
     	(0.9 * size(Imgs(jj).raw, 2) * white);
     ii = find(f_footrow, 1);  % first 90% white row
     
-    % If failed, instead look for black. 
+    % If failed, instead look for black (e.g., NRC footer).
     if isempty(ii)
         f_footrow = sum(Imgs(jj).raw, 2) == 0;
         ii = find(f_footrow, 1);  % first black row
@@ -182,52 +182,54 @@ for jj=1:length(Imgs)
 
         %-- Detecting magnification and/or pixel size ----------------%
         if license('test', 'video_and_image_blockset')  % check if toolbox for OCR is installed
-            o1 = ocr(footer);
+            
+            %-- OCR ------------------------------------------------------%
+            o1 = ocr(footer);  % first try, straight OCR on footer
+
+            % If failed, try again on binary image.
             if isempty(o1.Text)
                 o1 = ocr(footer > 0);
             end
 
-            Imgs(jj).ocr = o1;
+            % If text not as expected. Retry OCR. This time without scale bar.
+            if ~any(contains(o1.Text, {'nm', 'pm', 'um'}))
+                rp = regionprops(footer > 0);
+                ar = [rp.BoundingBox];
+                ar = reshape(ar', [4, length(ar)/4])';
+                ar = ar(:,3) ./ ar(:,4);  % arrive at aspect ratio
+                [~, ar_max] = max(ar);
+                rp_bb = floor(rp(ar_max).BoundingBox);
     
-            % Look for pixel size.
+                footer_nobar = footer;
+                footer_nobar(rp_bb(2):rp_bb(2)+rp_bb(4), ...
+                    rp_bb(1):rp_bb(1)+rp_bb(3)) = 0;  % remover scale bar before OCR
+
+                o1 = ocr(footer_nobar > 0);
+            end
+
+            Imgs(jj).ocr = o1;
+            
+            %-- Interpret OCR text ---------------------------------------%
+            % Look for pixel size directly.
             txts = {'nm/pix', 'nmlpix', 'nm/plx', 'nm/101x',...
                 'um/pix', 'umlpix','um/plx', 'um/101x',...
                 'pm/pix', 'pmlpix','pm/plx', 'pm/101x', ...
-                'nm', 'um'};
+                'nm', 'um', 'pm'};
             
+            % Loop through options. Flag if nm/um.
             for kk = 1:length(txts)
                 pixsize_end = strfind(o1.Text, txts(kk)) - 1;
                 if ~isempty(pixsize_end)
                     % Mark if unit read is not nm.
                     if or(contains(txts(kk), 'um'), ...
                             contains(txts(kk), 'pm'))
-                        f_nm = 0;
+                        fl_nm = 0;
                     end
                     break;
                 end
             end
             
-            % Alternative footer at NRC (black footer with scale bar).
-            % Scale bar code from below.
-            if isempty(pixsize_end)
-                footer = footer > 0;
-                rp = regionprops(footer);
-                ar = [rp.BoundingBox];
-                ar = reshape(ar', [4, length(ar)/4])';
-                ar = ar(:,3) ./ ar(:,4);  % arrive at aspect ratio
-                [~, ar_max] = max(ar); len = rp(ar_max).BoundingBox(3);  % pixel length of scale bar
-
[…]
