### 2024-01-20 8770102
Bug fix to EDM SBS
Fitting is now more robust.
+ Update to plotting in main_0 script
