### 2024-02-15 #5962 @tsipkens
I will give this some thought.

From: Kristen Thyng ***@***.***>
Sent: February 13, 2024 10:54 AM
Cc: Tim Sipkens ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: atems: Analysis tools for TEM images of carbonaceous particles (Issue #5962)

@tsipkens<https://github.com/tsipkens> you can see the difficulty in finding reviewers for this submission. Can you dig deeper into your extended network to suggest 5 more people?
