### 2024-02-16 6142b88
Improve scale recognition in load_imgs
