### 2024-02-16 9f8f7e0
Update imshow* functions
