### 2024-02-16 7086f19
Reduce size of border in imshow_binary
