+++ +tools/imshow_binary.m
@@ -28,11 +28,13 @@ if ~opts.f_outline
     
 else % if adding an outline
     img_edge = edge(img_binary,'sobel');
-    se = strel('disk',1);
-    img_dilated = imdilate(img_edge,se);
-        % use dilation to strengthen the aggregate's outline
-    i1 = uint8(~img_dilated).*t0;
-        % adds borders to labeled regions
+    
+    % Use dilation to strengthen the aggregate's outline.
+    se = strel('disk', 1);
+    img_dilated = or(img_binary - imdilate(img_binary, se), img_edge);
+    
+    % Adds borders to labeled regions.
+    i1 = uint8(~img_dilated) .* t0;
 end
 
 tools.imshow(i1); % show labelled image
