### 2024-02-16 7e08a02
Second update to boundary of imshow_binary
