### 2024-02-19 833e8d2
Add scale bar to some imshow
