### 2024-02-19 f9b7f06
Create aggcrop.m
