### 2024-02-19 c668198
Bug fixes etc. to scale display
