### 2024-02-19 529291f
Bug fix to overlay_scale
Dimensions were flipped, resulting in undesired behavior for small and non-square images.
