+++ +tools/overlay_scale.m
@@ -31,26 +31,26 @@ bar_length = bar_length1 / pixsize;  % in pixels
 
 % Properties for scale bar.
 margin = floor((size(I) .* 0.05));
-bar_height = max(margin(2) / 7, 1);
+bar_height = margin(2) / 5;
 font_specs = {'HorizontalAlignment', 'right', ...
     'VerticalAlignment', 'Bottom', ...
     'FontSize', 11, 'FontWeight', 'bold'};   % properties of text label
 
 hold on;
 rectangle('Position', ...
-    [size(I,1) - margin(1) - bar_length, ...
-     size(I,2) - margin(2), ...
+    [size(I,2) - margin(2) - bar_length, ...
+     size(I,1) - margin(1), ...
      bar_length, bar_height], ...
      'FaceColor', [0 0 0], ...
      'Curvature', 0.5);
 
 
 if bar_length1 > 1e3  % then use micron
-    text(size(I,1) - margin(1), size(I,2) - margin(2) - bar_height/5, ...
+    text(size(I,2) - margin(2), size(I,1) - margin(1) - bar_height/5, ...
         [num2str(bar_length1 ./ 1e3), ' um'], ...
         font_specs{:});
 else
-    text(size(I,1) - margin(1), size(I,2) - margin(2) - bar_height/5, ...
+    text(size(I,2) - margin(2), size(I,1) - margin(1) - bar_height/5, ...
         [num2str(bar_length1), ' nm'], ...
         font_specs{:});
 end
