### 2024-03-18 #6416 @jonbmartin
I have completed my review and I am satisfied with both the software toolbox and the package describing it, and believe they are appropriate for publication in JOSS. The MATLAB functions seem robust - I tried to add a few different types of noise to the images before segmentation and processing and got what seemed like satisfactory results every time.

The GUI interface (documented in main_0.m) built as a MATLAB app is very nice and clean, and a flexible addition to the automatic tools. However it is a little bit difficult to use. I found myself struggling to use the slider to specify 'threshold relative to Otsu' - allowing users to input a specific number than rely on a slider could help a bit. I also had some trouble with tools.load_imgs(), I believe as a linux user. I was able to run the command without arguments succesfully, but could never get it to succesfully accept a string input as the file path. Wasn't able to fully troubleshoot.

The paper is very well written, as is the documentation that was developed. I think that overall this is an excellent product and highly recommend publication.
