### 2024-03-21 #6416 @diazrenata
@jonbmartin thank you very much! @tsipkens, the two points raised (the somewhat finicky slider, and the question about running tools.load_imgs() with a string input) seem fairly minor, but do you have any thoughts in response?
