### 2024-04-03 e49530a
Fix README error
README contained an incorrect call to agg.seg(...). Should be agg.seg_kmeans(...). This matches main_b.
