### 2024-04-10 656b514
Amend load_imgs to also take a file name
+ Minor refactoring of spacing.
