### 2024-04-10 3fea642
Add Linux load limitation to README
