### 2024-04-10 #6416 @tsipkens
@diazrenata. Some changes have been made to the code in response to @jonbmartin's recommendations.

> The GUI interface (documented in main_0.m) built as a MATLAB app is very nice and clean, and a flexible addition to the automatic tools. However, it is a little bit difficult to use. I found myself struggling to use the slider to specify 'threshold relative to Otsu' - allowing users to input a specific number rather than rely on a slider could help a bit.

The GUI was amended accordingly, adding a numerical field in addition to the slider.

> I also had some trouble with tools.load_imgs(), I believe as a Linux user. I was able to run the command without arguments successfully, but could never get it to successfully accept a string input as the file path. Wasn't able to fully troubleshoot.

It is unclear as to the source of this issue. There is a reasonable chance this is related to the use of Linux. The CircleCI workflow still compiles, though this uses a weblink as an input. A note about potential issues for Linux users has been added to the README. The load_imgs function was also amended to take file names in addition to a folder, which could present another potential source for error.
