### 2024-04-11 #6416 @tytell
I have completed my review. As far as I can tell, the software does what it claims to do. My expertise is mainly in image analysis, and so I cannot address the specifics of analysis as it relates to soot or carbonaceous particles. The analysis is based on what seems to be a community need and seems based in existing literature. The paper is clear and well written and the scripts are documented and seem to run as described.

Here are a few notes and comments. I've added a few more minor issues to the repository.

1. Run `main_b.m`: Works

2. Run `tools.write_json(Aggs, 'test.json');`. Fails with

> Error using jsonencode
> Unable to encode objects of class logical as JSON-formatted text.
> Error in tools.write_json (line 9)
> t0 = jsonencode(var); % generate json text using built-in function

I think this is a Matlab version incompatibility. Here is my Matlab version info:

> MATLAB Version: 9.14.0.2254940 (R2023a) Update 2
> Operating System: macOS  Version: 12.6.9 Build: 21G726
> Java Version: Java 1.8.0_202-b08 with Oracle Corporation Java HotSpot(TM) 64-Bit Server VM mixed mode

3. Run `main_kmeans.m`. Works

4. Run `main_auto.m`. Works

5. Run `main_0.m`. Suggestions:
  - Change "more particles or add to existing particles" button to be labeled something like "Interactive tool to adjust thresholding"
  - Adjusting the "slider" does not seem to offer any feedback on how it is changing the threshold. That makes it very hard to judge whether the change is sufficient. I suggest overlying a color that shows the thresholded region.
  - I'm not sure if I'm doing it incorrectly, but every time I adjust the slider, I get a black image afterward, suggesting that the threshold is incorrect
  - There does not seem to be a way to go back in the steps. That would be helpful
  - Now I see that "Add (to) aggregates" button is sort of like a "start over" option. It is unclear what that actually does

6. Several of the functions (`main_kmeans` and the final section of `main_0` after `agg.analyze_binary`) occasionally crash Matlab on my computer. The program locks up entirely and I have to force quit. I can't determine why.

7. Run `img_binary = agg.seg_slider_orig(imgs)`. Works

8. README refers to `agg.seg_otsu_rb_orig(...)` and `agg.seg_otsu_rb(...)`. There aren't these functions in the agg directory. I think these should be `agg.seg_otsu_orig(...)` and `agg.seg_otsu(...)`.

9. README refers to `pp.kook*(...)`. I think this should be `pp.hough_kook*`.
