### 2024-05-07 591e7c7
Add header to write_json
This includes adding a note about a known issue with this function on Mac operating systems.
