### 2024-05-07 4932b28
Update dialog box text in agg.seg
