### 2024-05-07 9c79149
Bug fix to seg function
