### 2024-05-07 82270ca
Update README to note macOS/Linux limitations
