### 2024-05-07 2cc4396
Update README with more info for Linux
