### 2024-05-07 e5d4616
Bug fix to write_json
Correctly handle sparse logical `binary` field when encoding as JSON.
