+++ +tools/write_json.m
@@ -1,8 +1,6 @@
 
 % WRITE_JSON  Used to write structures for JSON files. 
-%  
-%  NOTE: There is a known issue on Mac systems where logicals cannot be
-%  written to the files. Omit the corresponding fields for those cases. 
+%  NOTE: Saving large aggregate structures can be slow.
 %  
 %  AUTHOR: Timothy Sipkens
 
@@ -10,6 +8,13 @@ function [t2,t0] = write_json(var, fname)
 
 fid = fopen(fname,'wt'); % open file, overwriting previous text
 
+% Binary field of Aggs structure is sparse logical. 
+% Convert of logical, as sparse structure cannot be stored. 
+if isfield(var, 'binary')
+    for ii=1:length(var)
+        var(ii).binary = full(var(ii).binary);
+    end
+end
 
 %-- Encode json ----------------------------------------------------------%
 t0 = jsonencode(var); % generate json text using built-in function
