### 2024-05-07 #6416 @tsipkens
Thanks to @tytell for the useful feedback. Brief responses to the review below:

1. The JSON encoding error has been resolved (related to encoding a sparse logical, which is not supported by JSON).

2. For main_0, the button for refining the threshold has been changed as suggested. Help buttons have also been added to the GUI (including why the thresholding may be failing), in addition to adding a text field for the slider related to the previous review. Color has been added to the overlay for the thresholding/slider step.

3. MATLAB locking up on occasion for some scripts could not be replicated and may be related to the use of EDM sizing on macOS. A note to this effect has been added to the README.

4. The reviewer is right, the README has been updated to contain correct references to the *seg_otsu* and *kook* methods.

All corresponding issues against the repository have been closed with comments.
