### 2024-05-28 0be7511
Fix type on paper.md
