### 2024-05-28 2db85b8
Update circleci config
Previous Ubuntu version deprecated.
