+++ .circleci/config.yml
@@ -6,7 +6,7 @@
 jobs:
   build:
     machine:
-      image: 'ubuntu-2004:202201-02'
+      image: default
     steps:
       - checkout
       - matlab/install
