### 2024-06-14 4275911
Improvements/fixes to getting footer
