### 2024-06-19 2da7347
Bug fix to load_imgs
Previous update removed some necessary code.
