### 2024-07-19 #6416 @diazrenata
@jonbmartin Thank you very much!

@tsipkens: I believe we are ready to move towards acceptance. Could you please (1) make a new release reflecting the changes that have occurred since review and (2) archive the project (e.g. on Zenodo or figshare)? When you're done, could you post the latest version number and archive DOI here?
