### 2024-07-19 #6416 @diazrenata
@tsipkens You can copy-paste the bullet pointed items in the above comment to create a checklist for yourself as we move towards acceptance!
