### 2024-07-23 #6416 @tsipkens
@diazrenata These tasks have been completed. Also note that the missing DOI corresponds to a thesis that, to the best of my knowledge, does not have an associated DOI.
- [x] Double check authors and affiliations (including ORCIDs). (**Noted that some authors do not have ORCIDs**)
- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. (**v1.3**)
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. (**10.5281/zenodo.12797363**)
